+++ lib/src/models/models.dart
@@ -1,1 +1,2 @@
 export 'video_quality_url.dart';
+export 'vimeo_video.dart';