+++ lib/tivy.dart
@@ -1,4 +1,5 @@
-/// Utility package for getting video information of YouTube and Vimeo video urls.
+/// Utility package for getting video quality urls for vimeo and youtube. 
 library tivy;
 
+export 'src/models/models.dart';
 export 'src/tivy.dart';