+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.0+1
 
-- Get `vimeo` video quality urls by url or id.
-- Get `vimeo` private video quality urls by url or id.
+- Get `vimeo` video information by url.
+- Get `vimeo` video quality urls by id or id.
+- Get `vimeo` private video quality urls by id.
 - Get `Youtube` video quality urls by url or id.